requirements.txt
python-docx==1.1.2
openai==1.58.1
pyyaml==6.0.2